// Frontend/MusicMarketBulgaria/src/app/ads/ad_enums/subCategories.enum.ts
// src/app/enums/sub-categories.enum.ts
export enum SubCategories {
    PERCUSSION = 'ударни',
    WOODWIND = 'духови',
    STRING = 'струнни',
    BRASS = 'медни',
    KEYBOARD = 'клавишни',
    STUDIO = 'студио',
    PA = 'озвучаване',
    OTHERS = 'други',
  }
  